BUG Fix cache file deletions for subsites and during publishing.
Split the delete function (unpublishPagesAndStaleCopies) into two
functions that separately handle deletions of regular and stale files.
Also provide a helper convertUrlsToPathMap which encapsulates some of
the complexity of making subsites publish into directiories.

The following two features are being fixed here:
- Make the deletions aware of subsites - files need to be deleted
from subdirectories if we are running in the subsites mode.
- Regular cache files need to be deleted on page modifications,
while the stale pages stay behind until the builder queue is executed.
This was originally working, but was broken with 2.0 refactor.