Added Translatable support
SiteTreeFullBuildEngine would only add pages with the default locale to the queue. This change checks if Translatable exists, if yes it disables the locale filter to return all urls.

I'm not certain this is the best way to do this, but it works as far as I can tell. Basically I just copied the code form 3 lines above for the Subsite Module.