FIX index files are being deleted when using domain based caching
When using domain based caching and a full cache build is triggered it
will generate index.html and index.php files correctly. However, when
clearing up old URLs it does a comparison of processed URLs with
actively published static files.

It tries to map the filepath to the url in
FilesystemPublisher->pathToURL() but doesn't trim the `/index`. When it
does a comparison, its checking `/index` against the actual URL `/`.
Anything published URL which doesn't match is deleted.